+++ StereoPlus2/ImGuiExtensions.hpp
@@ -185,5 +185,4 @@ namespace ImGui::Extensions {
         colors[ImGuiCol_NavWindowingDimBg]      = ImVec4(0.80f, 0.80f, 0.80f, 0.20f);
         colors[ImGuiCol_ModalWindowDimBg]       = ImVec4(0.80f, 0.80f, 0.80f, 0.35f);
     }
-
 }