+++ StereoPlus2/Input.hpp
@@ -18,6 +18,13 @@ class Input {
 			return lhs.code < rhs.code || lhs.code == rhs.code && lhs.type < rhs.type;
 		}
 	};
+	struct CombinationNode {
+		std::map<Key::Modifier, CombinationNode> children;
+		
+		// End keys and callbacks
+		std::map<Key::KeyPair, std::function<void()>, KeyPairComparator> callbacks;
+	};
+
 
 	// Continuous input is a state when there is
 	// input with delay in between.
@@ -45,6 +52,8 @@ class Input {
 		}
 	};
 
+	const Log Logger = Log::For<Input>();
+
 	glm::vec2 mouseOldPos;
 	glm::vec2 mouseNewPos;
 
@@ -57,6 +66,8 @@ class Input {
 	ContinuousInput continuousInputOneSecondDelay = ContinuousInput(1);
 	ContinuousInput continuousInputNoDelay = ContinuousInput(0);
 
+	CombinationNode combinations;
+
 	void FillAxes() {
 		mouseOldPos = mouseNewPos;
 		mouseNewPos = ImGui::GetMousePos();
@@ -66,34 +77,94 @@ class Input {
 			-IsPressed(Key::Left, useDiscreteMovement) + IsPressed(Key::Right, useDiscreteMovement),
 			IsPressed(Key::Up, useDiscreteMovement) + -IsPressed(Key::Down, useDiscreteMovement),
 			0);
-		NumpadAxe = glm::vec3(
-			-IsPressed(Key::N4, useDiscreteMovement) + IsPressed(Key::N6, useDiscreteMovement),
-			-IsPressed(Key::N2, useDiscreteMovement) + IsPressed(Key::N8, useDiscreteMovement),
-			-IsPressed(Key::N9, useDiscreteMovement) + IsPressed(Key::N1, useDiscreteMovement));
+		NumpadAxe = IsPressed(Key::Modifier::Shift)
+			? glm::vec3()
+			: glm::vec3(
+				-IsPressed(Key::N4, useDiscreteMovement) + IsPressed(Key::N6, useDiscreteMovement),
+				-IsPressed(Key::N2, useDiscreteMovement) + IsPressed(Key::N8, useDiscreteMovement),
+				-IsPressed(Key::N9, useDiscreteMovement) + IsPressed(Key::N1, useDiscreteMovement));
 
-		auto mouseMoveDirection = MouseMoveDirection();
-		MouseAxe = IsPressed(Key::MouseRight)
-			? glm::vec3(0, 0, -mouseMoveDirection.y)
-			: glm::vec3(mouseMoveDirection.x, -mouseMoveDirection.y, 0);
+		if (IsCustomRenderImageActive().Get()) {
+			auto mouseMoveDirection = MouseMoveDirection();
+			MouseAxe = IsPressed(Key::MouseRight)
+				? glm::vec3(0, 0, -mouseMoveDirection.y)
+				: glm::vec3(mouseMoveDirection.x, -mouseMoveDirection.y, 0);
+
+		}
+		else MouseAxe = glm::vec3();
+	}
+
+
+	static bool InsertCombination(
+		CombinationNode& node,
+		const std::vector<Key::Modifier>::const_iterator& modifiersCurrent,
+		const std::vector<Key::Modifier>::const_iterator& modifiersEnd,
+		const Key::KeyPair& endKey,
+		const std::function<void()>& callback) {
+		if (modifiersCurrent == modifiersEnd) {
+			if (node.children.find(*modifiersCurrent._Ptr) != node.children.end())
+				return false;
+
+			node.callbacks[endKey] = callback;
+			return true;
+		}
+
+		if (auto c = node.children.find(*modifiersCurrent._Ptr);
+			c != node.children.end())
+			return InsertCombination(c._Ptr->_Myval.second, modifiersCurrent + 1, modifiersEnd, endKey, callback);
+
+		node.children[*modifiersCurrent._Ptr] = CombinationNode();
+		return InsertCombination(node.children[*modifiersCurrent._Ptr], modifiersCurrent + 1, modifiersEnd, endKey, callback);
+	}
+
+	// Prevents combinations with letters or printable symbols
+	// being executed while the keyboard is captured by text input.
+	// Applied to Shift-modified and non-modified keys.
+	static bool IsAcceptableCombination(const int key, const bool ignoreCapturedKeys, const int level, const Key::Modifier& modifier) {
+		if (!ignoreCapturedKeys || key < GLFW_KEY_SPACE || key > GLFW_KEY_WORLD_2 && key < GLFW_KEY_KP_0 || key > GLFW_KEY_KP_EQUAL)
+			return true;
+
+		if (level == 0 || level == 1 && modifier == Key::Modifier::Shift)
+			return false;
+		
+		return true;
+	}
+
+	// Executes most complex combination if found.
+	static bool ExecuteFirstMatchingCombination(const CombinationNode& node, const bool ignoreCapturedKeys, const int level = 0, const Key::Modifier modifier = Key::Modifier::None) {
+		for (auto& [m, c] : node.children)
+			if (IsPressed(m) && ExecuteFirstMatchingCombination(c, false, level + 1, m))
+				return true;
+
+		for (auto& [k, c] : node.callbacks)
+			if (IsAcceptableCombination(k.code, ignoreCapturedKeys, level, modifier) && IsPressed(k, true)) {
+				c();
+				return true;
+			}
+
+		return false;
 	}
 
 public:
-	GLFWwindow* glWindow;
+	StaticProperty(GLFWwindow*, GLFWindow);
 	ImGuiIO* io;
 
 	glm::vec3 ArrowAxe;
 	glm::vec3 NumpadAxe;
 	glm::vec3 MouseAxe;
 
-
 	glm::vec3 movement;
 	std::map<size_t, std::function<void()>> handlers;
+
+	StaticProperty(bool, IsCustomRenderImageActive)
 
 	bool IsMoved() {
 		return movement != glm::vec3();
 	}
 
-	bool IsPressed(Key::KeyPair key, bool discreteInput = false) {
+	static bool IsPressed(Key::KeyPair key, bool discreteInput = false) {
+		auto* io = &ImGui::GetIO();
+
 		if (discreteInput)
 			return key.type == Key::Type::Mouse
 				? ImGui::IsMouseClicked(key.code, true)
@@ -103,7 +174,7 @@ public:
 			? io->MouseDown[key.code]
 			: io->KeysDown[key.code];
 	}
-	bool IsPressed(Key::Modifier key) {
+	static bool IsPressed(Key::Modifier key) {
 		auto* io = &ImGui::GetIO();
 
 		switch (key)
@@ -114,26 +185,10 @@ public:
 		}
 	}
 
-	bool IsDown(const Key::KeyPair& key) {
+	static bool IsDown(const Key::KeyPair& key) {
 		return key.type == Key::Type::Mouse
 			? ImGui::IsMouseClicked(key.code, true)
 			: ImGui::IsKeyPressed(key.code, true);
-	}
-	bool IsDown(const Key::Combination& keys) {
-		auto& io = ImGui::GetIO();
-
-		if (keys.modifiers.size() > 0) {
-			for (int i = 0; i < keys.modifiers.size() - 1; i++)
-				if (!IsPressed(keys.modifiers[i]) || keys.modifiers[i] == Key::Modifier::Shift && io.WantCaptureKeyboard)
-					return false;
-		}
-		else if (io.WantCaptureKeyboard)
-			return false;
-
-		if (!IsDown(keys.key))
-			return false;
-
-		return true;
 	}
 
 	bool IsUp(Key::KeyPair key) {
@@ -146,7 +201,7 @@ public:
 		return mouseNewPos;
 	}
 	glm::vec2 MouseMoveDirection() {
-		return glm::length(mouseNewPos - mouseOldPos) == 0 ? glm::vec2(0) : glm::normalize(mouseNewPos - mouseOldPos);
+		return glm::length(mouseNewPos - mouseOldPos) == 0 ? glm::vec2(0) : mouseNewPos - mouseOldPos;
 	}
 
 	// Does not include mouse movement nor scrolling.
@@ -163,26 +218,28 @@ public:
 	float MouseSpeed() {
 		return glm::length(mouseNewPos - mouseOldPos);
 	}
-	void SetMouseBoundlessMode(bool enable) {
-		if (enable == isMouseBoundlessMode)
-			return;
-
+	static void SetMouseBoundlessMode(bool enable) {
 		if (enable) {
-			if (isRawMouseMotionSupported)
-				glfwSetInputMode(glWindow, GLFW_RAW_MOUSE_MOTION, GLFW_TRUE);
+			if (glfwRawMouseMotionSupported())
+				glfwSetInputMode(GLFWindow().Get(), GLFW_RAW_MOUSE_MOTION, GLFW_TRUE);
 
-			glfwSetInputMode(glWindow, GLFW_CURSOR, GLFW_CURSOR_DISABLED);
+			glfwSetInputMode(GLFWindow().Get(), GLFW_CURSOR, GLFW_CURSOR_DISABLED);
 		}
 		else {
 			// We can disable raw mouse motion mode even if it's not supported
 			// so we don't bother with checking it.
-			glfwSetInputMode(glWindow, GLFW_RAW_MOUSE_MOTION, GLFW_FALSE);
+			glfwSetInputMode(GLFWindow().Get(), GLFW_RAW_MOUSE_MOTION, GLFW_FALSE);
 
-			glfwSetInputMode(glWindow, GLFW_CURSOR, GLFW_CURSOR_NORMAL);
+			glfwSetInputMode(GLFWindow().Get(), GLFW_CURSOR, GLFW_CURSOR_NORMAL);
 		}
-
-		isMouseBoundlessMode = enable;
 	}
+
+	void AddShortcut(const Key::Combination& combination, const std::function<void()>& callback) {
+		if (InsertCombination(combinations, combination.modifiers.begin(), combination.modifiers.end(), combination.key, callback))
+			return;
+		Logger.Error("Shortcut is already taken.");
+	}
+
 
 	void ProcessInput() {
 		movement = glm::vec3();
@@ -190,6 +247,11 @@ public:
 
 		continuousInputOneSecondDelay.Process(io->AnyKeyPressed);
 		continuousInputNoDelay.Process(io->AnyKeyPressed);
+
+		// Make sure printable characters don't trigger combinations
+		// while keyboard is captured by text input
+		if (io->AnyKeyPressed)
+			ExecuteFirstMatchingCombination(combinations, io->WantCaptureKeyboard);
 
 		// Handle OnInput actions
 		for (auto [id,handler] : handlers)
@@ -243,15 +305,13 @@ public:
 	}
 	void BindInputToMovement() {
 		AddHandler([i = input] {
-			bool isAltPressed = i->IsPressed(Key::Modifier::Alt);
+			// Enable or disable Mouse boundless mode 
+			bool mouseBoundlessMode = i->IsPressed(Key::Modifier::Alt) || i->IsPressed(Key::Modifier::Shift) || i->IsPressed(Key::Modifier::Control);
+			Input::SetMouseBoundlessMode(mouseBoundlessMode && Input::IsCustomRenderImageActive().Get());
 
-			// Enable or disable Mouose boundless mode 
-			// regardless of whether we move the cross or not.
-			i->SetMouseBoundlessMode(isAltPressed);
-
-			i->movement += i->MouseAxe * Settings().TranslationStep().Get();
-			i->movement += i->ArrowAxe * Settings().TranslationStep().Get();
-			i->movement += i->NumpadAxe * Settings().TranslationStep().Get();
+			i->movement += i->MouseAxe * Settings::MouseSensivity().Get();
+			i->movement += i->ArrowAxe;
+			i->movement += i->NumpadAxe;
 			});
 	}
 	void Cross() {